refactor(broker): delete duplicate broker core, unify governance floor onto exec-broker (reconciliation R4)
Removes the parallel scope core (src/broker/{scope,decide,needs}.ts) that
shadowed the canonical src/exec-broker/. The governance-floor scope check
(withGovernance.scopeNeeds) now reads the SAME signed source as the PreToolUse
gates — profileBrokerPolicy — via the new exec-broker/scope-needs.ts, using the
shared pure decisions (checkEgress/checkFsWrite) and policyFsRoots. kit doctor's
exec-broker posture row is likewise sourced from profileBrokerPolicy, preserving
the honest warn(unsigned)/fail(invalid/revoked/malformed) split via
verifyProfileSignature.

Behavior preserved (same denial messages, same fail-closed semantics). Leaves one
broker: src/exec-broker/* (canonical decisions + signed profile-policy provider),
with the gates, governance-floor scope check, kit doctor row, and host-extractor
(src/broker/extract.ts, kept) as the surfaces on top. Full suite 2698 pass / 0 fail.
